+++ Cargo.toml
@@ -20,7 +20,7 @@ rand_core = { version = "0.6", default-features = false }
 rand_chacha = "0.3"
 subtle = "2.6.1"
 halo2curves = { version = "0.8.0", features = ["bits", "derive_serde"] }
-generic-array = "1.1"
+generic-array = "1.2.0"
 num-bigint = { version = "0.4.6", features = ["serde", "rand"] }
 num-traits = "0.2.19"
 num-integer = "0.1.46"
@@ -28,9 +28,9 @@ serde = { version = "1.0.217", features = ["derive"] }
 bincode = "1.3"
 bitvec = "1.0"
 byteorder = "1.4.3"
-thiserror = "2.0.9"
+thiserror = "2.0.11"
 once_cell = "1.18.0"
-itertools = "0.13.0"
+itertools = "0.14.0"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { version = "0.2.15", default-features = false, features = ["js"] }
